pyproject.toml
[project]
name = "aws-bq-project"
version = "0.1.0"
description = "AWS to BigQuery data transfer project"
authors = [
    {name = "Sonali Cornelio",email = "sonali.cornelio@rea-group.com"}
]
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "google-cloud-bigquery (>=3.38.0,<4.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "boto3 (>=1.40.74,<2.0.0)",
    "dotenv (>=0.9.9,<0.10.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
